First Development of Arm Code
1. Needs PID tunning
2. Needs IDing
3. Needs encoder angles from the relative encoder
4. Needs peer review

How it works:
D-pad up, right, and down, while true passes an angle and a climbArmSubsystem into the Climb Command. The Climb Subsystem creates and configures the motors and encoders. This system also controls the motor output as well as the return angle. The Climb Command uses a wpilib's pid to send motor output values into the Climb Subsystem based off of the encoder's angle and the requested setpoint.