Motors operate based on BeamBreak;
Indexer motors will now stop spinning based on BreakBeam inputs, not a duration of time. New variables for each BeamBreak sensor which are used in conjunction with HasNote.

When a Note needs to be moved within the robot, a Direction will be passed to IndexerSubsystem.set() method. Depending on the direction (and HasNote), motors will be set to spin a certain direction and a "passing" variable will be set to true. If a "passing" variable is true, periodic() will check for the necessary stop condition (a specific BeamBreak reading) and stop all motors when it is met.

BACKWARD_INDEXER is now BACKWARD_INTAKE to match the DIRECTION_DESTINATION format,